split up show_editor.py
