+++ LightDrive/Workspace/Widgets/ShowEditor/show_editor.py
@@ -1,221 +1,15 @@
-from PySide6.QtWidgets import QMainWindow, QGraphicsView, QGraphicsScene, QGraphicsRectItem, QGraphicsItemGroup, \
-    QGraphicsPolygonItem, QGraphicsLineItem, QGraphicsTextItem, QGraphicsItem
-from PySide6.QtGui import QPen, QPolygonF, QWheelEvent, QPainter, QColor
-from PySide6.QtCore import Qt, QPointF, QRectF, QTimer, QElapsedTimer, QThread, Signal, QObject
+from Workspace.Widgets.ShowEditor.show_editor_general import TimingTickBar, Playhead
+from Workspace.Widgets.ShowEditor.show_editor_snippets import SnippetTrack, SnippetItem
+from Workspace.Widgets.ShowEditor.show_editor_audio import AudioTrack, WaveformItem, Markers, AudioLoaderWorker
+from PySide6.QtWidgets import QMainWindow, QGraphicsView, QGraphicsScene
+from PySide6.QtGui import QWheelEvent
+from PySide6.QtCore import Qt, QTimer, QElapsedTimer, QThread
 from tinytag import TinyTag
 import numpy as np
 import PySoundSphere
 import librosa
 import uuid
 import os
-
-class AudioLoaderWorker(QObject):
-    finished = Signal(np.ndarray, int)
-    error = Signal(str)
-
-    def __init__(self, audio_path):
-        super().__init__()
-        self.audio_path = audio_path
-
-    def run(self):
-        try:
-            y, sr = librosa.load(self.audio_path)
-            self.finished.emit(y, sr)
-        except Exception as e:
-            self.error.emit(str(e))
-
-class TimingTickBar(QGraphicsItemGroup):
-    def __init__(self, show_editor) -> None:
-        """
-        Create the timing tick bar
-        :param show_editor: The show editor
-        """
-        super().__init__()
-        self.show_editor = show_editor
-        self.ticks = []
-        self.labels = []
-
-        self.body = QGraphicsRectItem(0, 0, self.show_editor.track_length, 50)
-        self.body.setBrush(Qt.lightGray)
-        self.setOpacity(0.5)
-        self.addToGroup(self.body)
-
-        self.update_ticks()
-
-    def update_ticks(self) -> None:
-        """
-        Update the ticks, labels and body of the timing tick bar
-        :return: None
-        """
-        for tick in self.ticks:
-            self.removeFromGroup(tick)
-            self.scene().removeItem(tick)
-        self.ticks = []
-        for label in self.labels:
-            self.removeFromGroup(label)
-            self.scene().removeItem(label)
-        self.labels = []
-
-        for i in range(0, self.show_editor.track_length, 100):
-            x_pos = self.show_editor.virtual_frame_from_x_pos(i)
-            tick = QGraphicsLineItem(x_pos, 0, x_pos, 50)
-            tick.setPen(QPen(Qt.black, 1))
-            self.ticks.append(tick)
-            self.addToGroup(tick)
-            label = QGraphicsTextItem(str(i / 100) + "s")
-            label.setPos(x_pos, 0)
-            self.labels.append(label)
-            self.addToGroup(label)
-        self.body.setRect(0, 0, self.show_editor.track_length * self.show_editor.zoom, 50)
-
-class Playhead(QGraphicsItemGroup):
-    def __init__(self) -> None:
-        """
-        Create a playhead
-        :return: None
-        """
-        super().__init__()
-
-        # Create the parts of the playhead
-        self.body = QGraphicsRectItem(0, 0, 3, 100)
-        self.body.setBrush(Qt.red)
-        self.addToGroup(self.body)
-        self.head = QGraphicsPolygonItem(QPolygonF([QPointF(-8.5, -20), QPointF(11.5, -20), QPointF(1.5, 0)]))
-        self.head.setBrush(Qt.red)
-        self.addToGroup(self.head)
-        self.setZValue(1)
-
-class AudioTrack(QGraphicsRectItem):
-    def __init__(self, show_editor) -> None:
-        """
-        Create an audio track
-        :param show_editor: The show editor
-        """
-        super().__init__()
-        self.show_editor = show_editor
-
-        # Create the track
-        self.setRect(0, 0, self.show_editor.track_length, 100)
-        self.setBrush(Qt.lightGray)
-        self.setOpacity(0.25)
-
-    def update_width(self) -> None:
-        self.setRect(0, 0, self.show_editor.track_length * self.show_editor.zoom, 100)
-
-class WaveformItem(QGraphicsItem):
-    def __init__(self, waveform: np.ndarray, sample_rate: int, width: int, height: int, show_editor) -> None:
-        super().__init__()
-        self.waveform = waveform
-        self.sample_rate = sample_rate
-        self.width = width
-        self.height = height
-        self.show_editor = show_editor
-
-    def boundingRect(self) -> QRectF:  # noqa: N802
-        return QRectF(0, 0, self.width, self.height)
-
-    def paint(self, painter: QPainter, option, widget=None) -> None:
-        visible_rect = self.show_editor.mapToScene(self.show_editor.viewport().rect()).boundingRect()
-        visible_start = max(0, int(visible_rect.left()))
-        visible_end = min(int(self.width), int(visible_rect.right()))
-
-        color = QColor(33, 58, 211)
-        pen = QPen(color, 1)
-        painter.setPen(pen)
-        mid_y = self.height / 2
-        step = len(self.waveform) / self.width
-        for x in range(visible_start, visible_end):
-            start_idx = int(x * step)
-            end_idx = int((x + 1) * step)
-            segment = self.waveform[start_idx:end_idx]
-            max_amplitude = np.max(np.abs(segment))
-            y = mid_y - (max_amplitude * mid_y)
-            painter.drawLine(int(x), int(mid_y), int(x), int(y))
-
-    def update_width(self) -> None:
-        self.width = self.show_editor.track_length * self.show_editor.zoom
-
-class Markers(QGraphicsItem):
-    def __init__(self, show_editor, marker_times: np.ndarray, color, start_y: int, end_y: int) -> None:
-        super().__init__()
-        self.show_editor = show_editor
-        self.marker_times = marker_times
-        self.color = color
-        self.start_y = start_y + 100
-        self.end_y = end_y + 100
-        self.width = self.show_editor.track_length
-        self.height = 100
-
-    def boundingRect(self) -> QRectF:  # noqa: N802
-        return QRectF(0, 0, self.width, self.height)
-
-    def paint(self, painter: QPainter, option, widget=None) -> None:
-        pen = QPen(self.color, 1)
-        painter.setPen(pen)
-        for marker_time in self.marker_times:
-            x_pos = self.show_editor.virtual_frame_from_x_pos(marker_time * 100)
-            painter.drawLine(x_pos, self.start_y, x_pos, self.end_y)
-
-    def update_width(self) -> None:
-        self.width = self.show_editor.track_length * self.show_editor.zoom
-
-class SnippetTrack(QGraphicsRectItem):
-    def __init__(self, show_editor) -> None:
-        super().__init__()
-        self.show_editor = show_editor
-
-        # Create the track
-        self.setRect(0, 0, self.show_editor.track_length, 100)
-        self.setBrush(Qt.lightGray)
-        self.setOpacity(0.3)
-
-    def update_width(self) -> None:
-        self.setRect(0, 0, self.show_editor.track_length * self.show_editor.zoom, 100)
-
-class SnippetItem(QGraphicsItemGroup):
-    def __init__(self, show_editor, snippet_item_uuid: str, snippet_uuid: str, length: int = 250, frame: int = 0, track: int = 0) -> None:
-        super().__init__()
-        self.uuid = snippet_item_uuid
-        self.show_editor = show_editor
-        self.snippet_uuid = snippet_uuid
-        self.length = length
-        self.frame = frame
-        self.track = track
-
-        width = self.show_editor.virtual_frame_from_x_pos(self.length)
-        self.body = QGraphicsRectItem(0, 0, width, 100)
-        self.body.setBrush(Qt.green)
-        self.body.setOpacity(0.25)
-        self.addToGroup(self.body)
-
-        snippet_name = self.show_editor.window.snippet_manager.available_snippets[snippet_uuid].name
-        self.label = QGraphicsTextItem(snippet_name)
-        self.label.setPos(0, 0)
-        self.addToGroup(self.label)
-
-        self.setFlag(QGraphicsItem.ItemIsMovable)
-        self.setFlag(QGraphicsItem.ItemIsSelectable)
-        self.setFlag(QGraphicsItem.ItemSendsGeometryChanges)
-
-        x_pos = self.show_editor.x_pos_from_virtual_frame(self.frame)
-        y_pos = self.track * 100 + 50
-        self.setPos(x_pos, y_pos)
-
-    def itemChange(self, change, value):  # noqa: N802
-        if change == QGraphicsItem.ItemPositionChange:
-            # Keep in the middle of the tracks
-            track_number = round((value.y() - 50) / 100)
-            value.setY(track_number * 100 + 50)
-            if value.y() < 150: # Upper bounds
-                value.setY(150)
-            if value.x() < 0: # Left bounds
-                value.setX(0)
-            self.frame = self.show_editor.virtual_frame_from_x_pos(value.x())
-            if self.show_editor.show_snippet.added_snippets.get(self.uuid):
-                self.show_editor.show_snippet.added_snippets[self.uuid]["track"] = track_number
-                self.show_editor.show_snippet.added_snippets[self.uuid]["frame"] = self.frame
-
-        return super().itemChange(change, value)
 
 class ShowEditor(QGraphicsView):
     def __init__(self, window: QMainWindow, show_snippet) -> None:
